requirements.txt

httplib2
paho-mqtt